Use -Wno-pointer-sign to silence a lot of warnings.
We liberally assign between SQLCHAR * and char *, which generates a lot of
compiler warnings about different pointer signedness. Seems like a lost
cause to add casts everywhere (and it might not be good for robustness
anyway, because it might mask a bug where the datatype of one side is
changed to a struct or something in the future), so just silence the
warnings.