Use plain MSBuild project files for Windows builds.
This removes all the NMake files and PowerShell scripts, and modifies the
MSBuild project files formerly used with the PowerShell scripts so that they
can be used directly with MSBuild.exe. Also remove the .bat files used for
building the installers, and remove them with MSBuild targets.

MSBuild project files are now the only way to build on Windows. Assuming all
the dependencies are installed in standard locations, you can now do a full
build by simply typing "MSBuild" at the top-level in the source tree.